modified migration table

+++ database/migrations/2025_01_03_095422_create_tour_packages_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('package_name');
             $table->string('description');
-            $table->decimal('price', 10, 2);
+            $table->float('price', 2);
             $table->timestamps();
         });
     }
