+++ database/migrations/2025_01_03_095439_create_attractions_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('location');
             $table->string('description');
-            $table->decimal('price', 10, 2);
+            $table->float('price', 2);
             $table->timestamps();// Created and updated timestamps
         });
     }
